my-calendar-app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
mysql-connector-python
google-auth
google-auth-oauthlib
google-auth-httplib2
